requirements.txt
# === CORE SYSTEM ===
anyio==4.12.0
attrs==25.4.0
click==8.3.1
colorama==0.4.6
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
typing_extensions==4.15.0

# === API & SERVER ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
starlette>=0.38.0
sse-starlette==3.0.4
requests>=2.31.0
httpx==0.28.1
# watchfiles>=0.21.0
websockets>=12.0
pyzmq>=25.0.0

# === MCP PROTOCOL ===
mcp==1.25.0
fastmcp==2.14.1

# === MACHINE LEARNING (The Heavy Lifters) ===
# Using CPU versions for lighter images where possible
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
# torch handled in Dockerfile for CPU optimization
# torch>=2.0.0

# === DATA & STORAGE ===
supabase>=2.0.0
pyiceberg>=0.5.0
sqlite3-binary>=0.1.0
redis==5.0.1
chromadb>=0.4.0

# === UTILS ===
pyyaml>=6.0.0
python-dateutil>=2.8.2
pytz>=2024.1
psutil>=5.9.0
tqdm>=4.66.0

# === TELEMETRY ===
prometheus_client==0.23.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1

# === SECURITY ===
cryptography>=42.0.0
authlib>=1.3.0
pyjwt>=2.8.0
